Hotfix Deleted duplicate folders Changed resize behavior in MapHeader (to be improved for longer titles)
