+++ src/component/MapHeader/MapHeader.js
@@ -1,4 +1,4 @@
-import React, {useContext} from 'react'
+import React, { useContext, useState, useEffect, useCallback } from 'react'
 import { DownOutlined } from '@ant-design/icons';
 import { Button, Dropdown, Space } from 'antd';
 import "./MapHeader.css";
@@ -6,9 +6,19 @@ import { FlightsContext } from "../../container/FlightsProvider/FlightsProvider"
 
 
 const MapHeader = () => {
-  const {flights: items, currentFlight, switchCurrentFlight} = useContext(FlightsContext);
+  const { flights, currentFlight, switchCurrentFlight } = useContext(FlightsContext);
+
+  const items = flights.slice().reverse();
   
-  const handleMenuClick = ( {key}) => {
+  // Ant design key name fix
+  if (items.length > 0) {
+    items.forEach((flightObj, index) => {
+      flightObj['label'] = flightObj['date'].slice(0, 10);
+      flightObj['key'] = items.length - 1 - index;
+    })
+  }
+
+  const handleMenuClick = ({ key }) => {
     switchCurrentFlight(key);
   };
 
@@ -16,30 +26,32 @@ const MapHeader = () => {
     items,
     onClick: handleMenuClick,
     selectable: true,
-        defaultSelectedKeys: ['0'],
+    defaultSelectedKeys: [`${items.length - 1}`],
+    //defaultSelectedKeys: [`0`],
   };
-  
 
   return (
     <div className="switch-flights">
-        <span>
-            Lot nr. {currentFlight.key+1}
+      <div>
+        <span className="switch-flights-title">
+          {!isNaN(currentFlight.key) ? currentFlight.title : null} 
         </span>
-        <Space wrap className="switch-flights-button-wrapper">
-            <Dropdown 
-              menu={menuProps}
-              trigger={['click']}
-              getPopupContainer={() => document.querySelector('.switch-flights')}
-              overlayClassName="switch-flights-dropdown"
-            >
-                <Button>
-                    <Space>
-                    {currentFlight.label}
-                    <DownOutlined />
-                    </Space>
-                </Button>
-            </Dropdown>
-        </Space>
+      </div>
+      <Space wrap className="switch-flights-button-wrapper">
+        <Dropdown
+          menu={menuProps}
+          trigger={['click']}
+          getPopupContainer={() => document.querySelector('.switch-flights')}
+          overlayClassName="switch-flights-dropdown"
+        >
+          <Button className="switch-button">
+            <Space>
+              {currentFlight.label}
+              <DownOutlined />
+            </Space>
+          </Button>
+        </Dropdown>
+      </Space>
     </div>
   )
 }
